Week9-Gen_AI/requirements.txt
langchain-openai        
langchain               
python-dotenv       
langchain_community     
langchain_core         
streamlit            
pypdf          
bs4                       
arxiv               
pymupdf                   
wikipedia           
duckduckgo-search
langchain-text-splitters 
chromadb         
langchain_huggingface 
sentence-transformers
faiss-cpu                
langchain_chroma    
duckdb             
pandas                
openai             
langchain-groq       
numpy                 
mysql-connector-pyth
SQLAlchemy         
validators==0.28.1
youtube_transcript_api 
unstructured           
numexpr                  
huggingface_hub
cassio
datasets
duckduckgo-search
torch
